% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\docType{class}
\name{A-class}
\alias{A-class}
\alias{rfb_A-class}
\alias{rb_A-class}
\alias{chr_A-class}
\title{A-class}
\description{
An S4 class to represent component Ay (the last advice or reference catch) 
of the rfb, rb, and chr rules.

The classes \code{rfb_A}, \code{rb_A}, and \code{chr_A} inherit from 
\code{A} and their only difference is that the slot \code{hcr}
is set to the corresponding catch rule name ('rfb', 'rb', or 'chr').
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of component Ay (reference catch)}

\item{\code{value_landings}}{Optional. The landings corresponding to \code{value}.}

\item{\code{value_discards}}{Optional. The discards corresponding to \code{value}.}

\item{\code{value_catch}}{Optional. The total catch corresponding to \code{value}. May differ from \code{value} if discard survival is considered.}

\item{\code{hcr}}{The harvest control rule (hcr) for which Ay is used. One of 'rfb', 'rb', or 'chr'.}

\item{\code{data}}{Time series of historical catches and/or advice}

\item{\code{avg_years}}{Number of years for calculating average catch}

\item{\code{basis}}{Basis of Ay. Either "advice" for using previous advice or "average catch" when based on average of historical catch}

\item{\code{advice_metric}}{Advice metric, 'catch' or 'landings'.}

\item{\code{discard_survival}}{Optional. Discard survival (%). Can be used to show the reference catch (or advice) in the form of dead catch (or advice).}
}}

