% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A.R
\name{A}
\alias{A}
\alias{A,numeric-method}
\alias{A,data.frame-method}
\alias{A,A-method}
\alias{rfb_A}
\alias{rfb_A,ANY-method}
\alias{rb_A}
\alias{rb_A,ANY-method}
\alias{chr_A}
\alias{chr_A,ANY-method}
\title{rfb/rb/chr rule - component Ay (reference catch or advice)}
\usage{
A(
  object,
  value,
  units,
  hcr,
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

rfb_A(
  object,
  value,
  units,
  hcr = "rfb",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "advice",
  ...
)

rb_A(
  object,
  value,
  units,
  hcr = "rb",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

chr_A(
  object,
  value,
  units,
  hcr = "chr",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)
}
\arguments{
\item{object}{The reference catch. See details}

\item{units}{[Optional] The units of the reference catch, e.g. "tonnes".}

\item{hcr}{[Optional] The harvest control rule (hcr) for which the multiplier is used. One of 'rfb', 'rb', or 'chr'.}

\item{data}{[Internal] Data used for calculating reference catch.}

\item{avg_years}{[Optional] Number of years for calculating average catch or vector years to use}

\item{basis}{[Optional] Basis of Ay. Either "advice" for using the previous advice or "average catch" when based on an average of historical catch}

\item{advice_metric}{Advice metric, e.g. catch or landings.}

\item{...}{Additional arguments (Not used)}
}
\value{
An object of class \code{A}
}
\description{
This function defines the reference catch (last advice or average of 
historical catches) for the rfb, rb, and chr rules.
}
\details{
The function accepts as its first argument (`object`):
- a single value representing a reference catch, e.g. the previous catch advice
- a vector of historical values which are used to calculate the average catch
- a data.frame with columns 'year' and either of 'advice', 'catch', 'landings'
- an object of class `A`

\code{rfb_A()}, \code{rb_A()}, and \code{chr_A()} are aliases for
\code{A()} in which the \code{hcr} argument is already set to 
'rfb', 'rb', or 'chr'.

The reference catch is set following ICES (2022).

Usually, the reference catch is the previous advised catch. Alternatively,
if the rfb/rb/chr rule is applied the first time, it can be based on an
average of historical catches.
}
\examples{
# data.frame with advice/catch history
df <- data.frame(year = 2018:2022,
                 advice = c(3257, 3648, 2721, 2177, 1742),
                 catch = c(2513, 2091, 1887, 1614, NA),
                 landings = c(1880, 1725, 1373, 1403, NA))
# get reference catch (advice)
A <- A(object = df, basis = "advice", units = "tonnes", advice_metric = "catch")
advice(A)
# use average catch
A(object = df, basis = "average catch", avg_years = 3, units = "tonnes", advice_metric = "catch")

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.
}
\keyword{internal}
