% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/comp_m.R
\name{advice}
\alias{advice}
\alias{advice,comp_m-method}
\alias{advice,rfb_m-method}
\alias{advice,rb_m-method}
\alias{advice,chr_m-method}
\title{ICES advice sheet-style output table}
\usage{
advice(object)
}
\arguments{
\item{object}{A component of any of the empirical harvest control rules or the the output from applying the rule.}
}
\value{
NULL. A table is displayed in the R terminal.
}
\description{
This function returns an advice sheet-style table for the empirical harvest 
control rules. The argument passed to the function can either be a single
component of any rules (components r, f, b, ...) or the output from applying
any of the rules (rfb, rb, chr).
}
\examples{
# multiplier for the rfb rule with known k
# advice(rfb_m(k = 0.1))

}
