% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\alias{b,data.frame-method}
\alias{b,I-method}
\alias{b,numeric-method}
\alias{b,b-method}
\alias{rfb_b}
\alias{rfb_b,ANY-method}
\alias{rb_b}
\alias{rb_b,ANY-method}
\alias{chr_b}
\alias{chr_b,ANY-method}
\title{rb/rfb/chr rule - component b (biomass safeguard)}
\usage{
b(object, idx_value, Itrigger, Iloss, w, yr_ref, lag, n_yrs, units, hcr, ...)

rfb_b(
  object,
  idx_value,
  Itrigger,
  Iloss,
  w,
  yr_ref,
  lag,
  n_yrs,
  units,
  hcr = "rfb",
  ...
)

rb_b(
  object,
  idx_value,
  Itrigger,
  Iloss,
  w,
  yr_ref,
  lag,
  n_yrs,
  units,
  hcr = "rb",
  ...
)

chr_b(
  object,
  idx_value,
  Itrigger,
  Iloss,
  w,
  yr_ref,
  lag,
  n_yrs,
  units,
  hcr = "chr",
  ...
)
}
\arguments{
\item{object}{The biomass index. Can be a \code{data.frame} with columns 'data' and 'index'.}

\item{idx_value}{Optional. The current index value. Only used if no index time series is supplied.}

\item{Itrigger}{Optional. The index trigger value below which the biomass safeguard reduces the catch advice.}

\item{Iloss}{Optional. The lowest index value, can be used to calculate \code{Itrigger}.}

\item{w}{Optional. The index trigger buffer (multiplier) to link \code{Itrigger} to \code{Iloss}. Defaults to \code{w=1.4}.}

\item{yr_ref}{Optional. If supplied, this specifies the year in the biomass index which is used as \code{Iloss} and \code{Itrigger} is calculated from this value.}

\item{lag}{Optional. Time lag between the last index year and the last year to be used. By default, the last index year is used (\code{lag=0}).}

\item{n_yrs}{Optional. The number of years used in the index. By default, only the last index value is used (\code{n_yrs=1}).}

\item{units}{Optional. The units of the biomass index, e.g. 'kg/hr'. Only used for plotting.}

\item{hcr}{Optional. One of 'rfb', 'rb', or 'chr'.}

\item{...}{Additional arguments. Not used.}
}
\value{
An object of class \code{b} with the value of the biomass 
safeguard
}
\description{
This function calculates component b (the biomass safeguard) of the rb, rfb,
and chr rule. The index needs to be a biomass index without age structure.
}
\details{
The biomass safeguard compares the last index value (\out{<i>I</i>}) to an index trigger value (\out{<i>I</i><sub>trigger</sub>}). If the current index value is below the trigger, the biomass safeguard reduces the catch advice:

\out{<i>b</i> = min{1, <i>I</i> / <i>I</i><sub>trigger</sub>}}


, where \out{<i>I</i><sub>trigger</sub>} is usually derived from the lowest observed biomass index value (\out{<i>I</i><sub>loss</sub>}) as:

\out{<i>I</i><sub>trigger</sub> = w * <i>I</i><sub>loss</sub>}

with 

\out{<i>w</i> = 1.4}

See ICES (2022) for the full definition definition.

Usually, this method is used by providing only a biomass index, e.g. 
as a \code{data.frame}. The method uses this index, searches for the lowest
index value (\code{Iloss}), multiplies this value by the index trigger
buffer (\code{w}) to get the index trigger value (\code{Itrigger}). 
The last index value in the time series is then compared to \code{Itrigger}
and if the index value is below, the biomass safeguard reduces the catch
advice.

The biomass safeguard is identical in the rfb, rb, and chr rules. 
\code{rfb_b()}, \code{rb_b()} and \code{chr_b()} are aliases for 
\code{b()} with identical arguments and functionality.
}
\section{Warning}{

Please note that \out{<i>I</i><sub>trigger</sub>} should only be defined once the first time the empirical harvest control rule is applied. In the following years, the same value should be used for \out{<i>I</i><sub>trigger</sub>}.
For application in ICES, do not change the defaults unless the change is supported by stock-specific simulations.
}

\examples{
# If the value of the biomass safeguard is known
b(1)

# First application of the biomass safeguard
# Use a data.frame with index values
df_idx <- data.frame(year = 2017:2021,
                     index = c(1.33, 1.13, 0.84, 0.60, 1.03))
b <- b(df_idx)
b
advice(b)

# plot
plot(b(df_idx, units = "kg/hr"))

# Use of the biomass safeguard in a following year without updating Itrigger
df_idx <- data.frame(year = 2017:2022,
                     index = c(1.33, 1.13, 0.84, 0.60, 1.03, 0.5))
b(df_idx, yr_ref = 2020)

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2022. Exploring a relative harvest rate strategy for moderately data-limited fisheries management. ICES Journal of Marine Science, 79: 1730--1741. \url{https://doi.org/10.1093/icesjms/fsac103}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Application of explicit precautionary principles in data-limited fisheries management. ICES Journal of Marine Science, 78: 2931--2942. \url{https://doi.org/10.1093/icesjms/fsab169}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Using a genetic algorithm to optimize a data-limited catch rule. ICES Journal of Marine Science, 78: 1311--1323. \url{https://doi.org/10.1093/icesjms/fsab018}.

Fischer, S. H., De Oliveira, J. A. A., and Kell, L. T. 2020. Linking the performance of a data-limited empirical catch rule to life-history traits. ICES Journal of Marine Science, 77: 1914--1926. \url{https://doi.org/10.1093/icesjms/fsaa054}.
}
