% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\docType{class}
\name{chr-class}
\alias{chr-class}
\title{An S4 class to represent the chr rule.}
\description{
This class contains the components of the chr rule 
(\code{\link{I}}, \code{\link{F}}, \code{b}, \code{m}).
}
\section{Slots}{

\describe{
\item{\code{advice}}{The value of the catch advice.}

\item{\code{advice_landings}}{Landings corresponding to the catch advice.}

\item{\code{advice_discards}}{Discards corresponding to the catch advice.}

\item{\code{advice_discards_dead}}{Dead discards corresponding to the catch advice.}

\item{\code{advice_discards_surviving}}{Surviving discards corresponding to the catch advice.}

\item{\code{advice_uncapped}}{The value of the catch advice without the uncertainty cap.}

\item{\code{units}}{The unit (e.g. tonnes) of the catch advice.}

\item{\code{advice_metric}}{The advice metric, 'catch' or 'landings'.}

\item{\code{frequency}}{The advice frequency (annual/biennial).}

\item{\code{years}}{The years for which the advice is valid.}

\item{\code{A}}{The reference catch (previous catch advice).}

\item{\code{I}}{Component I (the biomass index value).}

\item{\code{F}}{Component F (the relative harvest rate target).}

\item{\code{b}}{Component b (the biomass safeguard).}

\item{\code{m}}{Component m (the multiplier).}

\item{\code{cap}}{Uncertainty cap (stability clause, restricts changes in advice).}

\item{\code{cap_lower}}{Maximum allowed reduction in advice in \%, e.g. -30.}

\item{\code{cap_upper}}{Maximum allowed increase in advice in \%, e.g. 20.}

\item{\code{change}}{Change in advice compared to previous advice.}

\item{\code{change_uncapped}}{Change in advice compared to previous advice before application of the uncertainty cap.}

\item{\code{discard_rate}}{Discard rate (\%).}

\item{\code{discard_survival}}{Discard survival (\%).}
}}

