% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr.R
\name{chr}
\alias{chr}
\alias{chr,missing,A,I,F,b,m-method}
\alias{chr,missing,numeric,numeric,numeric,numeric,numeric-method}
\alias{chr,chr,missing,missing,missing,missing,missing-method}
\alias{chr,chr,A,I,F,b,m-method}
\title{chr rule}
\usage{
chr(
  object = new("chr"),
  A = object@A,
  I = object@I,
  F = object@F,
  b = object@b,
  m = object@m,
  cap = "conditional",
  cap_upper = 20,
  cap_lower = -30,
  years,
  frequency = "annual",
  discard_rate = NA,
  ...
)
}
\arguments{
\item{object}{Optional. An object of class \code{chr}.}

\item{A}{\code{\link{A}}. The reference catch (previous catch advice). Required for calculating change in advice and for the application of the stability clause.}

\item{I}{\code{\link{I}}. The biomass index value.}

\item{F}{\code{\link{F}}. The harvest rate target.}

\item{b}{\code{\link{b}}. The biomass safeguard.}

\item{m}{\code{\link{m}}. The multiplier.}

\item{cap}{The uncertainty cap (stability clause). Defaults to 'conditional', i.e. it is only considered when b=1.}

\item{cap_upper}{Optional. \code{numeric}. The maximum allowed increase in the catch advice in \%. Default to +20.}

\item{cap_lower}{Optional. \code{numeric}. The maximum allowed decrease in the catch advice in \%. Default to -20.}

\item{years}{Optional. \code{numeric}. The years for which the advice should be given.}

\item{frequency}{Optional. The frequency of the advice ('annual'/'biennial'/'triennial'). Defaults to 'annual'.}

\item{discard_rate}{Optional. \code{numeric}. The discard rate for the advice. If provided, advice values for catch and landings are given.}

\item{...}{Additional parameters. Not used.}
}
\value{
An object of class \code{chr}.
}
\description{
This function applies the chr rule.
}
\details{
This function applies the chr rule following the ICES technical guidelines (ICES, 2022). The function requires the elements of the chr rule: \out{<i>I</i>} (the biomass index, see \code{\link{chr_I}}), \out{<i>F</i><sub>MSYproxy</sub>} (the target harvest rate, see \code{\link{F}}), \out{<i>b</i>} (the biomass safeguard, see \code{\link{chr_b}}) and \out{<i>m</i>} (the multiplier, see \code{\link{chr_m}}). The catch advice is then calculated as

\out{<i>A</i><sub>y+1</sub> = <i>I</i> * <i>F</i><sub>MSYproxy</sub> * <i>b</i> * <i>m</i>}

restricted by the stability clause relative to \out{<i>A</i><sub>y</sub>}.
See the help files of the components for their definition (\code{\link{chr_I}}, \code{\link{F}}, \code{\link{chr_b}}, \code{\link{chr_m}})
}
\section{Warning}{

For application in ICES, do not change the default parameters (frequency,
stability clause, etc) unless the changes are supported by case-specific
simulations.
}

\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2022. Exploring a relative harvest rate strategy for moderately data-limited fisheries management. ICES Journal of Marine Science, 79: 1730--1741. \url{https://doi.org/10.1093/icesjms/fsac103}.
}
