% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/f.R
\name{indicator}
\alias{indicator}
\alias{indicator,f-method}
\title{indicator}
\usage{
indicator(object)

\S4method{indicator}{f}(object)
}
\arguments{
\item{object}{The object from which the indicator is requested.}
}
\description{
This function returns the indicator value(s) for the empirical
harvest control rules, e.g. the length indicator.
}
