% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_A.R
\docType{class}
\name{comp_A-class}
\alias{comp_A-class}
\alias{rfb_A-class}
\alias{rb_A-class}
\alias{chr_A-class}
\title{comp_A-class}
\description{
An S4 class to represent component Ay (the last advice or reference catch) 
of the rfb, rb, and chr rules.

The classes \code{rfb_A}, \code{rb_A}, and \code{chr_A} inherit from 
\code{comp_A} and their only difference is that the slot \code{hcr}
is set to the corresponding catch rule name ('rfb', 'rb', or 'chr').
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of component Ay (reference catch)}

\item{\code{hcr}}{The harvest control rule (hcr) for which Ay is used. One of 'rfb', 'rb', or 'chr'.}

\item{\code{data}}{Time series of historical catches and/or advice}

\item{\code{avg_years}}{Number of years for calculating average catch}

\item{\code{basis}}{Basis of Ay. Either "advice" for using previous advice or "average catch" when based on average of historical catch}

\item{\code{advice_metric}}{Advice metric, 'catch' or 'landings'.}
}}

