% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_A.R
\name{comp_Ay}
\alias{comp_Ay}
\alias{comp_Ay,numeric-method}
\alias{comp_Ay,data.frame-method}
\alias{comp_Ay,comp_Ay-method}
\alias{rfb_Ay}
\alias{rfb_Ay,ANY-method}
\alias{rb_Ay}
\alias{rb_Ay,ANY-method}
\alias{chr_Ay}
\alias{chr_Ay,ANY-method}
\title{rfb/rb/chr rule - component Ay (reference catch or advice)}
\usage{
comp_Ay(
  object,
  value,
  units,
  catch_rule,
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

\S4method{comp_Ay}{numeric}(
  object,
  value,
  units,
  catch_rule,
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

\S4method{comp_Ay}{data.frame}(
  object,
  value,
  units,
  catch_rule,
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

\S4method{comp_Ay}{comp_Ay}(
  object,
  value,
  units,
  catch_rule,
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

rfb_Ay(
  object,
  value,
  units,
  catch_rule = "rfb",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "advice",
  ...
)

\S4method{rfb_Ay}{ANY}(
  object,
  value,
  units,
  catch_rule = "rfb",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "advice",
  ...
)

rb_Ay(
  object,
  value,
  units,
  catch_rule = "rb",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

\S4method{rb_Ay}{ANY}(
  object,
  value,
  units,
  catch_rule = "rb",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

chr_Ay(
  object,
  value,
  units,
  catch_rule = "chr",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)

\S4method{chr_Ay}{ANY}(
  object,
  value,
  units,
  catch_rule = "chr",
  data,
  avg_years,
  basis = "advice",
  advice_metric = "catch",
  ...
)
}
\arguments{
\item{object}{The reference catch. See details}

\item{units}{[Optional] The units of the reference catch, e.g. "tonnes".}

\item{catch_rule}{[Optional] The catch rule for which the multiplier is used. One of 'rfb', 'rb', or 'chr'.}

\item{data}{[Internal] Data used for calculating reference catch.}

\item{avg_years}{[Optional] Number of years for calculating average catch or vector years to use}

\item{basis}{[Optional] Basis of Ay. Either "advice" for using the previous advice or "average catch" when based on an average of historical catch}

\item{advice_metric}{Advice metric, e.g. catch or landings.}

\item{...}{Additional arguments (Not used)}
}
\value{
An object of class \code{comp_Ay}
}
\description{
This function defines the reference catch (last advice or average of 
historical catches) for the rfb, rb, and chr rules.
}
\details{
The function accepts as its first argument (`object`):
- a single value representing a reference catch, e.g. the previous catch advice
- a vector of historical values which are used to calculate the average catch
- a data.frame with columns 'year' and either of 'advice', 'catch', 'landings'
- an object of class `comp_Ay`

\code{rfb_Ay()}, \code{rb_Ay()}, and \code{chr_Ay()} are aliases for
\code{comp_Ay()} in which the \code{catch_rule} argument is already set to 
'rfb', 'rb', or 'chr'.

The reference catch is set following ICES (2022).

Usually, the reference catch is the previous advised catch. Alternatively,
if the rfb/rb/chr rule is applied the first time, it can be based on an
average of historical catches.
}
\examples{

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.
}
\keyword{internal}
