% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_b.R
\docType{class}
\name{comp_b-class}
\alias{comp_b-class}
\alias{rfb_b-class}
\alias{rb_b-class}
\alias{chr_b-class}
\title{comp_b-class}
\description{
An S4 class to represent component b of the rfb/rb/chr rules.

This class (\code{comp_b}) stores the input for component b (the biomass
safeguard) as well as the resulting b value. 

The classes \code{rfb_b}, \code{rb_b}, and \code{chr_b} inherit from 
\code{comp_b} and their only difference is that the slot \code{catch_rule}
is set to the corresponding catch rule name ('rfb', 'rb', or 'chr').
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of component b}

\item{\code{idx_value}}{Index value that is compared to Itrigger.}

\item{\code{Itrigger}}{The index trigger value below which the advice is reduced. Usually calculated as \code{Itrigger=Iloss*w}.}

\item{\code{Iloss}}{The lowest observed index value. Can be used as the basis for \code{Itrigger}.}

\item{\code{w}}{Index trigger buffer. Connects \code{Itrigger} to \code{Iloss}.}

\item{\code{yr_ref}}{Reference year on which Itrigger is based.}

\item{\code{yr_last}}{Last data year of the biomass index. The index value in this year is compared to \code{Itrigger}.}

\item{\code{n0}}{Time lag between the last index year and the last year to be used.}

\item{\code{idx}}{\code{data.frame}. A \code{data.frame} with the index values.}

\item{\code{units}}{\code{character}. The units of the biomass index, e.g. 'kg/hr'.}

\item{\code{catch_rule}}{\code{character}. The catch rule for which the biomass safeguard is used. One of 'rfb', 'rb', or 'chr'.}
}}

