% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_b.R
\name{comp_b}
\alias{comp_b}
\alias{comp_b,FLQuant-method}
\alias{comp_b,data.frame-method}
\alias{comp_b,numeric-method}
\alias{comp_b,comp_b-method}
\alias{rfb_b}
\alias{rb_b}
\alias{chr_b}
\title{rb/rfb/chr rule - component b (biomass safeguard)}
\usage{
comp_b(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)

\S4method{comp_b}{FLQuant}(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)

\S4method{comp_b}{data.frame}(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)

\S4method{comp_b}{numeric}(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)

\S4method{comp_b}{comp_b}(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)

comp_b(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)

comp_b(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)

comp_b(object, idx_value, Itrigger, Iloss, w, yr_ref, n0, units, ...)
}
\arguments{
\item{object}{The biomass index. Can be a \code{data.frame} with columns 'data' and 'index' or an \code{FLQuant} object defined by \code{FLCore}.}

\item{idx_value}{Optional. The current index value. Only used if no index time series is supplied.}

\item{Itrigger}{Optional. The index trigger value below which the biomass safeguard reduces the catch advice.}

\item{Iloss}{Optional. The lowest index value, can be used to calculate \code{Itrigger}.}

\item{w}{Optional. The index trigger buffer (multiplier) to link \code{Itrigger} to \code{Iloss}. Defaults to \code{w=1.4}.}

\item{yr_ref}{Optional. If supplied, this specifies the year in the biomass index which is used as \code{Iloss} and \code{Itrigger} is calculated from this value.}

\item{n0}{Optional. Time lag between the last index year and the last year to be used. By default, the last index year is used (\code{n0=0})}

\item{units}{Optional. The units of the biomass index, e.g. 'kg/hr'. Only used for plotting.}

\item{...}{Additional arguments. Not used.}

\item{catch_rule}{Optional. One of 'rfb', 'rb', or 'chr'.}
}
\value{
An object of class \code{comp_b} with the value of the biomass 
safeguard
}
\description{
This function calculates component b (the biomass safeguard) of the rb, rfb,
and chr rule. The index needs to be a biomass index without age structure.
}
\details{
The biomass safeguard compares the last index value to an index trigger 
value. If the current index value is below the trigger, the biomass safeguard
reduces the catch advice. See ICES (2022) for the definition.

Usually, this method is used by providing only a biomass index, e.g. 
as a \code{data.frame}. The method uses this index, searches for the lowest
index value (\code{Iloss}), multiplies this value by the index trigger
buffer (\code{w}) to get the index trigger value (\code{Itrigger}). 
The last index value in the time series is then compared to \code{Itrigger}
and if the index value is below, the biomass safeguard reduces the catch
advice.

The biomass safeguard is identical in the rfb, rb, and chr rules. 
\code{rfb_b()}, \code{rb_b()} and \code{chr_b()} are aliases for 
\code{comp_b()} with identical arguments and functionality.
}
\section{Warning}{

For application in ICES, do not change the defaults unless the change is supported by stock-specific simulations.
}

\examples{
# If the value of the biomass safeguard is known
comp_b(1)

# Use a data.frame with index values
df_idx <- data.frame(year = 2017:2021,
                     index = c(1.33, 1.13, 0.84, 0.60, 1.03))
comp_b(df_idx)

# plot
plot(comp_b(df_idx, units = "kg/hr"))

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2022. Exploring a relative harvest rate strategy for moderately data-limited fisheries management. ICES Journal of Marine Science, 79: 1730--1741. \url{https://doi.org/10.1093/icesjms/fsac103}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Application of explicit precautionary principles in data-limited fisheries management. ICES Journal of Marine Science, 78: 2931--2942. \url{https://doi.org/10.1093/icesjms/fsab169}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Using a genetic algorithm to optimize a data-limited catch rule. ICES Journal of Marine Science, 78: 1311--1323. \url{https://doi.org/10.1093/icesjms/fsab018}.

Fischer, S. H., De Oliveira, J. A. A., and Kell, L. T. 2020. Linking the performance of a data-limited empirical catch rule to life-history traits. ICES Journal of Marine Science, 77: 1914--1926. \url{https://doi.org/10.1093/icesjms/fsaa054}.
}
