% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_f.R
\docType{class}
\name{comp_f-class}
\alias{comp_f-class}
\alias{rfb_f-class}
\title{comp_f-class}
\description{
An S4 class to represent component f of the rfb rule.

This class (\code{comp_f}) stores the input for component f (the length 
indicator as well as the resulting f value.
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of component f}

\item{\code{indicator}}{Length indicator time series}

\item{\code{yr_last}}{\code{numeric}. The last year with data.}

\item{\code{years}}{Years with data.}

\item{\code{Lmean}}{Mean catch length.}

\item{\code{Lref}}{Reference catch length.}

\item{\code{units}}{\code{character}. The units of the biomass index, e.g. 'kg/hr'.}

\item{\code{hcr}}{\code{factor}. The harvest control rule (hcr) for which component f is used (rfb).}
}}

