% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_I.R
\docType{class}
\name{comp_I-class}
\alias{comp_I-class}
\alias{chr_I-class}
\title{comp_I-class}
\description{
An S4 class to represent component I (the current biomass 
index value) of the chr rule.

This class (\code{comp_I}) stores the input for component I
as well as the resulting I value.
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of component I}

\item{\code{lag}}{The time lag (in years) between the last available index value and the value to be used.}

\item{\code{n_years}}{\code{numeric}. The number of years used for the index value.}

\item{\code{idx}}{\code{data.frame}. A \code{data.frame} with the index values.}

\item{\code{yr_last}}{\code{numeric}. The last year with index data.}

\item{\code{units}}{\code{character}. The units of the biomass index, e.g. 'kg/hr'.}

\item{\code{catch_rule}}{\code{character}. The catch rule for which the index is used. Only applicable to 'chr'.}
}}

