% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_m.R
\name{comp_m}
\alias{comp_m}
\alias{comp_m,numeric-method}
\alias{comp_m,comp_m-method}
\alias{comp_m,missing-method}
\alias{rfb_m}
\alias{rfb_m,ANY-method}
\alias{rfb_m,missing-method}
\alias{rb_m}
\alias{rb_m,ANY-method}
\alias{rb_m,missing-method}
\alias{chr_m}
\alias{chr_m,ANY-method}
\alias{chr_m,missing-method}
\title{rfb/rb/chr rule - component m (multiplier)}
\usage{
comp_m(object, catch_rule, k, ...)

rfb_m(object, value, catch_rule = "rfb", k, ...)

rb_m(object, value, catch_rule = "rb", k, ...)

chr_m(object, value, catch_rule = "chr", k, ...)
}
\arguments{
\item{object}{Optional. A multiplier m value, if known, or an existing \code{comp_m} object.}

\item{catch_rule}{The catch rule for which the multiplier is used. One of 'rfb', 'rb', or 'chr'.}

\item{k}{Optional. The von Bertalanffy k parameter (individual growth rate, unit: 1/year).}
}
\value{
An object of class \code{comp_m}
}
\description{
This function returns the default multiplier for the rfb, rb, and chr rules.
}
\details{
\code{rfb_m()}, \code{rb_m()}, and \code{chr_m()} are aliases for
\code{comp_m()} in which the \code{catch_rule} argument is already set to 
'rfb', 'rb', or 'chr'.

The multiplier is set following ICES (2022).

For the rfb rule, the multiplier is set depending on the von Bertalanffy 
parameter k (individual growth rate, units: 1/year; ICES, 2022). 
For species where 
k is below 0.2/year, the multiplier is set to m=0.95. For species where 
k is at or above 0.2/year, but below 0.32/year, the multiplier is set to
m=0.90. If the rfb rule is applied and k is unknown, the more precautionary
multiplier of m=0.90 is used.

For the rb rule, the multiplier is set to m=0.50 (ICES, 2022).

For the chr rule, the multiplier is set to m=0.50 (ICES, 2022).
}
\section{Warning}{

For application in ICES, do not change the default multiplier unless the
 change is supported by stock-specific simulations.
}

\examples{
# rfb rule with known k
rfb_m(k = 0.1) # 0.95
comp_m(catch_rule = "rfb", k = 0.1) # 0.95
rfb_m(k = 0.25) # 0.90
comp_m(catch_rule = "rfb", k = 0.25) # 0.90
# rfb rule with unknown k
rfb_m() # 0.90
comp_m(catch_rule = "rfb") # 0.90

# rb rule
rb_m() # 0.5
comp_m(catch_rule = "rb") # 0.5

# chr rule
chr_m() # 0.5
comp_m(catch_rule = "chr") # 0.5

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.
}
