% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\alias{f,missing,Lmean,Lref-method}
\alias{f,f,missing,missing-method}
\alias{f,numeric,missing,missing-method}
\alias{rfb_f}
\alias{rfb_f,missing,ANY,ANY-method}
\alias{rfb_f,ANY,missing,missing-method}
\title{rfb rule - component f (fishing pressure proxy, length indicator)}
\usage{
f(object, Lmean, Lref, n0 = 0, units, hcr, ...)
}
\arguments{
\item{object}{Optional. An object of class \code{f}.}

\item{Lmean}{The mean catch length. Either a \code{data.frame} with columns
'year' and 'Lmean' or an object of class \code{Lmean}.}

\item{Lref}{The reference length. Either a \code{numeric} with the value or 
an object of class \code{Lref}.}

\item{n0}{Time lag between the last indicator year and the last year to be used. Defaults to 0.}

\item{units}{Optional. The units of the length dat, e.g. 'cm'. Only used for plotting.}

\item{hcr}{Optional. Defaults to 'rfb'.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{f} with the length indicator value(s).
}
\description{
This function calculates component f (the fishing pressure proxy, derived from a length indicator ) of the rfb rule.
}
\details{
The value is calculated by comparing the mean catch length (above length of first capture Lc) to a reference length.

\code{rfb_f()} is an alias for \code{f()} with identical arguments and functionality.
}
\section{Note}{

The reference length Lref should be kept constant for all years unless there 
is a substantial changes in the fishery or fishery selectivity.
}

\examples{
# use ple7e example data
data(ple7e_length)
# calculate (pooled) length at first capture first
lc <- Lc(data = ple7e_length, pool = 2017:2021)
# calculate mean catch length
lmean <- Lmean(data = ple7e_length, Lc = lc, units = "mm")
# reference length
lref <- Lref(Lc = 264, Linf = 585)
# calculate component f
f <- f(Lmean = lmean, Lref = lref, units = "mm")
f
advice(f)
plot(f)

}
\references{
ICES. 2025. ICES Guidelines - Advice rules for stocks in category 2 and 3. Version 3. ICES Guidelines and Policies - Advice Technical Guidelines. 31 pp. \url{https://doi.org/10.17895/ices.pub.28506179}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Application of explicit precautionary principles in data-limited fisheries management. ICES Journal of Marine Science, 78: 2931--2942. \url{https://doi.org/10.1093/icesjms/fsab169}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Using a genetic algorithm to optimize a data-limited catch rule. ICES Journal of Marine Science, 78: 1311--1323. \url{https://doi.org/10.1093/icesjms/fsab018}.

Fischer, S. H., De Oliveira, J. A. A., and Kell, L. T. 2020. Linking the performance of a data-limited empirical catch rule to life-history traits. ICES Journal of Marine Science, 77: 1914--1926. \url{https://doi.org/10.1093/icesjms/fsaa054}.
}
