% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_hr.R
\docType{class}
\name{Ftarget-class}
\alias{Ftarget-class}
\title{F}
\description{
An S4 class to represent component F (the target harvest rate) of the chr rule.

This class (\code{F}) stores the input for the target harvest rate (if any) as well as the resulting target harvest rate.
}
\section{Slots}{

\describe{
\item{\code{value}}{The target harvest rate value.}

\item{\code{metric}}{The metric for the harvest rate (e.g. catch, landings, or
dead catch)}

\item{\code{data}}{\code{data.frame}. The data (harvest rates) used for calculating the target harvest rate.}

\item{\code{yr_ref}}{\code{numeric}. The years from which data are used.}

\item{\code{units}}{\code{character}. The units of the harvest rate.}

\item{\code{HR}}{\code{HR}. The harvest rate input data.}

\item{\code{indicator}}{\code{F}. The indicator used to select years of the harvest rate.}

\item{\code{hcr}}{\code{character}. The harvest control rule (hcr) for which the index is used. Only applicable to 'chr'.}

\item{\code{MSE}}{\code{logical}. Is the harvest rate a generic value or was it calculated with stock-specific simulations (MSE)? Defaults to \code{FALSE}.}

\item{\code{multiplier}}{\code{numeric}. Optional. Multiplier to adjust the target harvest rate. Only applicable if \code{MSE=TRUE}.}
}}

