% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_F_hr.R
\docType{class}
\name{HR-class}
\alias{HR-class}
\title{HR-class}
\description{
An S4 class to represent the harvest rate (hr) of the chr rule.

This class (\code{HR}) stores the input for the harvest rate 
(catch, landings) as well as the resulting harvest rates.
}
\section{Slots}{

\describe{
\item{\code{value}}{The values of the harvest rate time series.}

\item{\code{data}}{\code{data.frame}. The input data (catch and index values)}

\item{\code{units}}{\code{character}. The units of the harvest rate.}

\item{\code{units_catch}}{\code{character}. The units of the catch.}

\item{\code{units_index}}{\code{character}. The units of the index.}

\item{\code{hcr}}{\code{character}. The harvest control rule (hcr) for which the index is used. Only applicable to 'chr'.}
}}

