% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I.R
\name{I}
\alias{I}
\alias{I,data.frame-method}
\alias{I,vector-method}
\alias{I,I-method}
\alias{chr_I}
\alias{chr_I,ANY-method}
\title{chr rule - component I (biomass index value)}
\usage{
I(object, lag = 0, n_yrs = 1, units, hcr = "chr", ...)

chr_I(object, lag = 0, n_yrs = 1, units, hcr = "chr", ...)
}
\arguments{
\item{object}{The biomass index. Can be a \code{data.frame} with columns 'data' and 'index', a vector, or a single value.}

\item{lag}{Optional. The time lag (in years) between the last available index value and the value to be used. Defaults to 0 (the last value is used).}

\item{n_yrs}{Optional. The number of years if an average index value is used. Defaults to 1 (use last year's value only).}

\item{units}{Optional. The units of the biomass index, e.g. 'kg/hr'. Only used for plotting.}

\item{hcr}{Optional. Should be 'chr'.}

\item{...}{Additional arguments. Not used.}
}
\value{
An object of class \code{I} with the value of the biomass 
index
}
\description{
This function calculates component I (the last biomass index value) of the chr rule. The index needs to be a biomass index without age structure.
}
\details{
See ICES (2022) for the full definition definition.

Usually, this method is used by providing only a biomass index, e.g. as a \code{data.frame}. The method uses this index, and takes the last index value.

\code{chr_I()} is an alias for \code{I()} with identical arguments and functionality.
}
\section{Warning}{

For application in ICES, do not change the defaults unless the change is supported by stock-specific simulations.
}

\examples{
# Use a data.frame with index values
df_idx <- data.frame(year = 2017:2021,
                     index = c(1.33, 1.13, 0.84, 0.60, 1.03))
I <- I(df_idx)
I
advice(I)


# If only the value of the last biomass index is known
I(1)

# ple7e example data
data(ple7e_idx)
I <- I(ple7e_idx)
advice(I)
plot(I)

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2022. Exploring a relative harvest rate strategy for moderately data-limited fisheries management. ICES Journal of Marine Science, 79: 1730--1741. \url{https://doi.org/10.1093/icesjms/fsac103}.
}
