% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{inverse_indicator}
\alias{inverse_indicator}
\title{Return the inverse indicator for component f of the chr rule.}
\usage{
inverse_indicator(object)
}
\arguments{
\item{object}{An object of class \code{f}.}
}
\value{
A \code{data.frame} with the inverse length indicator value(s).
}
\description{
Return the inverse indicator for component f of the chr rule.
}
