% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_data.R
\docType{class}
\name{Lc-class}
\alias{Lc-class}
\title{An S4 class to represent length at first capture}
\description{
This class (\code{Lc}) stores the input data (length frequencies) for the
length at first capture and the resulting length at first capture value(s).
}
\section{Slots}{

\describe{
\item{\code{value}}{The length at first capture value(s)}

\item{\code{summary}}{A summary of the length at first capture value(s)}

\item{\code{years}}{The years used in the calculation of length at first capture}

\item{\code{pooled}}{Are length data from several years combined (pooled)?}

\item{\code{averaged}}{Are annual values from several years average?}

\item{\code{units}}{The units for length data (e.g. cm)}

\item{\code{data}}{The data (length frequencies) used in the calculation}
}}

