% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_data.R
\name{Lc}
\alias{Lc}
\alias{Lc,numeric-method}
\alias{Lc,Lc-method}
\alias{Lc,data.frame-method}
\title{Length at first capture Lc}
\usage{
Lc(
  data,
  pool = FALSE,
  average = FALSE,
  lmin,
  lmax,
  lstep,
  rounding = floor,
  units,
  ...
)

\S4method{Lc}{numeric}(
  data,
  pool = FALSE,
  average = FALSE,
  lmin,
  lmax,
  lstep,
  rounding = floor,
  units,
  ...
)

\S4method{Lc}{Lc}(
  data,
  pool = FALSE,
  average = FALSE,
  lmin,
  lmax,
  lstep,
  rounding = floor,
  units,
  ...
)

\S4method{Lc}{data.frame}(
  data,
  pool = FALSE,
  average = FALSE,
  lmin,
  lmax,
  lstep,
  rounding = floor,
  units,
  ...
)
}
\arguments{
\item{data}{The input data with the length distribution. (see details below)}

\item{pool}{Pool data from several years in the calculation?
\code{TRUE}/\code{FALSE} or a vector specifying years to use.}

\item{average}{Calculate Lc as the average of several annual values?
\code{TRUE}/\code{FALSE} or a vector specifying years to use.}

\item{lmin}{Optional. Smallest length class to use. If defined, length
classes below \code{lmin} are ignored.}

\item{lmax}{Optional. Largest length class to use. If defined, length
classes above \code{lmax} are ignored.}

\item{lstep}{Optional. Size of length classes. Allows combining length
classes into larger length classes. (see details below)}

\item{rounding}{Optional. The method used to round length classes when using
\code{lstep}. Defaults to \code{floor}, can also be
\code{ceiling} or \code{round}.}

\item{units}{Units of length data, e.g. "cm".}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{Lc}
}
\description{
This function calculates length at first capture Lc from a length frequency
distribution for use with the rfb/chr rules.
}
\details{
Length at first capture Lc is defined as the first length class, in which the
numbers of fish is at or above the mode of the distribution (the length class
with the highest number of fish).

The length distribution is passed to \code{Lc} with the argument \code{data}.
\code{data} is ideally a \code{data.frame} with columns "year", "length", and
"numbers".

The argument \code{pool} allows the pooling of length data from several years
in the estimation of Lc. If set to \code{FALSE} (default), Lc will be
calculated for each year, if set to \code{TRUE}, all years will be combined.
Alternatively, a vector of years can be provided and only these years will be
used.

The argument \code{average} allows the averaging of annual Lc values over
several years. If set to \code{FALSE} (default), Lc will be
calculated for each year, if set to \code{TRUE}, Lc values from all years
will be averaged. Alternatively, a vector of years can be provided and only
these years will be used for the average.

The optional arguments \code{lmin} and \code{lmax} allow removing of length
classes outside this range. \code{lstep} can be used to combine the length
into broader length classes. This can be useful if data are noisy with
several local minima/maxima and to smooth the length distribution.
}
\section{Note}{

For application with the rfb or chr rule, Lc should be set once in the first
year of the implementation. At every subsequent application of the method,
the previous Lc should then be kept unless a substantial change happened
(e.g. because of changed in the fishery or fishery selectivity).

data, pool = FALSE, lmin, lmax, lstep, rounding = floor
}

\section{Warning}{

For application in ICES, do not change the defaults (\code{n0}, \code{n1}, \code{n2}) unless the change is supported by stock-specific simulations.
}

\examples{

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.
}
