% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot,Lc,ANY-method}
\alias{plot,Lc,ANY-method}
\alias{plot,Lmean,ANY-method}
\title{Plot length frequencies}
\usage{
\S4method{plot}{Lc,ANY}(x, y, y_label, ...)

\S4method{plot}{Lmean,ANY}(x, y, y_label, ...)
}
\arguments{
\item{x}{An object of class \code{Lc}, \code{Lmean}, ...}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{gg}/\code{ggplot} with the plot.
Can be manipulated with the usual ggplot2 commands, e.g. \code{ylim()}.
}
\description{
A convenience function for plotting length frequencies and length reference
points.
}
\examples{

}
