% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_data.R
\docType{class}
\name{Lmean-class}
\alias{Lmean-class}
\title{An S4 class to represent mean catch length}
\description{
This class (\code{Lmean}) stores the input data (length frequencies) for the
mean catch length and the resulting mean catch length(s), above the length
at first capture (Lc).
}
\section{Slots}{

\describe{
\item{\code{value}}{The mean catch length value(s)}

\item{\code{summary}}{A summary of the mean catch length value(s)}

\item{\code{years}}{The years for which mean catch length is calculated}

\item{\code{Lc}}{The length at first capture}

\item{\code{include_Lc}}{Include Lc in the calculation of the mean length? (default: \code{TRUE})}

\item{\code{units}}{The units for length data (e.g. cm)}

\item{\code{data}}{The data (length frequencies) used in the calculation}
}}

