% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_data.R
\name{Lmean}
\alias{Lmean}
\alias{Lmean,numeric,missing-method}
\alias{Lmean,Lmean,missing-method}
\alias{Lmean,data.frame,missing-method}
\alias{Lmean,data.frame,data.frame-method}
\alias{Lmean,data.frame,Lc-method}
\title{Mean catch length Lmean}
\usage{
Lmean(data, Lc, lmin, lmax, lstep, rounding = floor, units, ...)

\S4method{Lmean}{numeric,missing}(data, Lc, lmin, lmax, lstep, rounding = floor, units, ...)

\S4method{Lmean}{Lmean,missing}(data, Lc, lmin, lmax, lstep, rounding = floor, units, ...)

\S4method{Lmean}{data.frame,missing}(data, Lc, lmin, lmax, lstep, rounding = floor, units, ...)

\S4method{Lmean}{data.frame,data.frame}(data, Lc, lmin, lmax, lstep, rounding = floor, units, ...)

\S4method{Lmean}{data.frame,Lc}(data, Lc, lmin, lmax, lstep, rounding = floor, units, ...)
}
\arguments{
\item{data}{The input data with the length distribution. (see details below)}

\item{Lc}{Length of first capture. Either a single value used for all years
or a data.frame with columns "year" and "Lc".}

\item{lmin}{Optional. Smallest length class to use. If defined, length
classes below \code{lmin} are ignored.}

\item{lmax}{Optional. Largest length class to use. If defined, length
classes above \code{lmax} are ignored.}

\item{lstep}{Optional. Size of length classes. Allows combining length
classes into larger length classes. (see details below)}

\item{rounding}{Optional. The method used to round length classes when using
\code{lstep}. Defaults to \code{floor}, can also be
\code{ceiling} or \code{round}.}

\item{units}{Units of length data, e.g. "cm".}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{Lc}
}
\description{
This function calculates the mean catch length above the length of first 
capture.
}
\details{
The mean catch length is calculated as the mean length of fish in the catch
which are above the length of first capture (Lc). The mean catch length is
essentially the mean of lengths, weighted by the number of fish in the length
classes.

The length distribution is passed to \code{Lmean} with the argument 
\code{data}. \code{data} is ideally a \code{data.frame} with columns "year", 
"length", and "numbers". An optional column, "catch_category", can be 
included to distinguish between categories such as "landings" and 
"discards".

The mean catch length calculation only considers length classes above the 
length of first capture \code{Lc}. \code{Lc} can be provided as a single 
value (recommended) or annual values can be provided with a \code{data.frame}
with columns "year" and "Lc". Generally, it is recommended to use a single 
value for Lc and only change it if there are substantial changes between 
years.

If \code{Lc} is not provided, the input \code{data} is passed to \code{Lc()}
to estimate Lc.

The optional arguments \code{lmin} and \code{lmax} allow removing of length
classes outside this range. \code{lstep} can be used to combine the length
into broader length classes. This can be useful if data are noisy with
several local minima/maxima and to smooth the length distribution.
}
\examples{

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.
}
