% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_data.R
\docType{class}
\name{Lref-class}
\alias{Lref-class}
\title{An S4 class to represent the reference catch length}
\arguments{
\item{basis}{The basis for the calculation, defaults to "LF=M".}

\item{Lc}{The length at first capture.}

\item{Linf}{The asymptotic length from a von Bertalanffy growth model.}

\item{Mk}{The ratio of natural mortality M to von Bertalanffy k.}

\item{gamma}{Links fishing mortality F to natural mortality M
to set the proxy for MSY (see details).}

\item{theta}{Alternative option to link M and k.}

\item{years}{Years, if annual values provided.}
}
\description{
This class (\code{Lref}) stores the value of the reference catch length, 
e.g. LF=M.
}
\section{Slots}{

\describe{
\item{\code{value}}{The reference catch length.}

\item{\code{units}}{The units for length data (e.g. cm)}
}}

