% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_data.R
\name{Lref}
\alias{Lref}
\title{Reference catch length}
\usage{
Lref(basis = "LF=M", Lc, Linf, Mk = 1.5, gamma = 1, theta = 1/Mk, units, ...)
}
\arguments{
\item{basis}{The basis for the calculation, defaults to "LF=M".}

\item{Lc}{The length at first capture.}

\item{Linf}{The asymptotic length from a von Bertalanffy growth model.}

\item{Mk}{The ratio of natural mortality M to von Bertalanffy k.}

\item{gamma}{Links fishing mortality F to natural mortality M
to set the proxy for MSY (see details).}

\item{theta}{Alternative option to link M and k.}

\item{units}{The units for length data (e.g. cm)}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{Lref}
}
\description{
This function calculates the reference catch length.
}
\details{
The default is to calculate the MSY proxy reference length following 
Beverton & Holt (1957) and as derived by Jardim et al. (2015):

\code{(theta * Linf + Lc * (gamma + 1)) / (theta + gamma + 1)}

where \code{Linf} is the asymptotic length of a von Bertalanffy growth model,
\code{Lc} the length of first capture, 
theta links von Bertalanffy individual growth parameter k and natural
mortality (M) through \code{k = theta * M} and gamma links fishing mortality 
F to M through \code{F = gamma * M}. The default reference length calculation
assumes \code{theta = 2/3}, i.e. that \code{M/k = 1.5} and that 
\code{gamma = 1}, i.e. that \code{F = M} can be used as a proxy for MSY.
The ratio M/k can be set directly with the argument \code{Mk} or indirectly
with \code{theta} which defaults to \code{1/Mk}.
}
\section{Warning}{
 Changing the default parameters is discouraged. 
Any change to the default parameters should be well justified.

The reference length is usually set once the first time the rfb rule is 
applied and should then be kept constant unless there a substantial changes
in the fishery or fishery selectivity.
}

\examples{
# calculate MSY proxy LF=M
Lref(Lc = 26.4, Linf = 58.5)

}
\references{
Beverton, R. J. H., and Holt, S. J. 1957. On the Dynamics of Exploited Fish Populations. Fishery Investigation Series 2. HMSO for Ministry of Agriculture, Fisheries and Food, London. 533 pp.
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.
Jardim, E., Azevedo, M., and Brites, N. M. 2015. Harvest control rules for data limited stocks using length-based reference points and survey biomass indices. Fisheries Research, 171: 12–19. \url{https://doi.org/10.1016/j.fishres.2014.11.013}.
}
