% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\docType{class}
\name{m-class}
\alias{m-class}
\alias{rfb_m-class}
\alias{rb_m-class}
\alias{chr_m-class}
\title{m-class}
\description{
An S4 class to represent component m (the multiplier) of the rfb, rb, and 
chr rules.

The classes \code{rfb_m}, \code{rb_m}, and \code{chr_m} inherit from 
\code{m} and their only difference is that the slot \code{hcr}
is set to the corresponding catch rule name ('rfb', 'rb', or 'chr').
}
\section{Slots}{

\describe{
\item{\code{value}}{The value of component m}

\item{\code{hcr}}{The harvest control rule (hcr) for which the multiplier is used. One of 'rfb', 'rb', or 'chr'.}

\item{\code{k}}{Optional. The von Bertalanffy k parameter (individual growth rate, unit: 1/year).}
}}

