% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\alias{m,numeric-method}
\alias{m,m-method}
\alias{m,missing-method}
\alias{rfb_m}
\alias{rfb_m,ANY-method}
\alias{rfb_m,missing-method}
\alias{rb_m}
\alias{rb_m,ANY-method}
\alias{rb_m,missing-method}
\alias{chr_m}
\alias{chr_m,ANY-method}
\alias{chr_m,missing-method}
\title{rfb/rb/chr rule - component m (multiplier)}
\usage{
m(object, hcr, k, MSE, ...)

rfb_m(object, hcr = "rfb", k, MSE, ...)

rb_m(object, hcr = "rb", k, MSE, ...)

chr_m(object, hcr = "chr", k, MSE, ...)
}
\arguments{
\item{object}{Optional. A multiplier m value, if known, or an existing \code{m} object.}

\item{hcr}{The harvest control rule (hcr) for which the multiplier is used. One of 'rfb', 'rb', or 'chr'.}

\item{k}{Optional. The von Bertalanffy k parameter (individual growth rate, unit: 1/year).}

\item{MSE}{Optional. \code{logical}. Default multiplier or multiplier derived from stock specific MSE?}

\item{...}{Additional arguments. Not used.}
}
\value{
An object of class \code{m}
}
\description{
This function returns the default multiplier for the rfb, rb, and chr rules.
}
\details{
\code{rfb_m()}, \code{rb_m()}, and \code{chr_m()} are aliases for
\code{m()} in which the \code{hcr} argument is already set to 
'rfb', 'rb', or 'chr'.

The multiplier is set following ICES (2025).

For the rfb rule, the multiplier is set depending on the von Bertalanffy 
parameter k (individual growth rate, units: 1/year; ICES, 2022). 
For species where 
k is below 0.2/year, the multiplier is set to m=0.95. For species where 
k is at or above 0.2/year, but below 0.32/year, the multiplier is set to
m=0.90. If the rfb rule is applied and k is unknown, the more precautionary
multiplier of m=0.90 is used.

For the rb rule, the multiplier is set to m=0.50 (ICES, 2022).

For the chr rule, the multiplier is set to m=0.50 (ICES, 2022).
}
\section{Warning}{

For application in ICES, do not change the default multiplier unless the
 change is supported by stock-specific simulations.
}

\examples{
# rfb rule with known k
rfb_m(k = 0.1) # 0.95
m(hcr = "rfb", k = 0.1) # 0.95
rfb_m(k = 0.25) # 0.90
m(hcr = "rfb", k = 0.25) # 0.90
# rfb rule with unknown k
rfb_m() # 0.90
m(hcr = "rfb") # 0.90

# rb rule
rb_m() # 0.5
m(hcr = "rb") # 0.5

# chr rule
chr_m() # 0.5
m(hcr = "chr") # 0.5

}
\references{
ICES. 2025. ICES Guidelines - Advice rules for stocks in category 2 and 3. Version 3. ICES Guidelines and Policies - Advice Technical Guidelines. 31 pp. \url{https://doi.org/10.17895/ices.pub.28506179}.
}
