% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ple7e}
\alias{ple7e}
\alias{ple7e_catch}
\alias{ple7e_f}
\alias{ple7e_f2}
\alias{ple7e_hr}
\alias{ple7e_idx}
\alias{ple7e_length}
\title{Western English Channel plaice example data}
\format{
An object of class \code{data.frame} with 36 rows and 5 columns.

An object of class \code{comp_f} of length 1.

An object of class \code{f} of length 1.

An object of class \code{HR} of length 1.

An object of class \code{data.frame} with 19 rows and 2 columns.

An object of class \code{data.frame} with 777 rows and 4 columns.
}
\source{
\url{https://github.com/ices-taf}
}
\usage{
ple7e_catch

ple7e_f

ple7e_f2

ple7e_hr

ple7e_idx

ple7e_length
}
\description{
This data set contains 2022 example data from Western English Channel plaice (ple.27.7e) to illustrate the application of the ICES category 3 empirical harvest control rules.
}
\keyword{datasets}
