% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rb.R
\name{rb}
\alias{rb}
\alias{rb,missing,A,r,b,m-method}
\alias{rb,rb,missing,missing,missing,missing-method}
\alias{rb,rb,A,r,b,m-method}
\title{rb rule}
\usage{
rb(
  object,
  A,
  r,
  b,
  m,
  cap = "conditional",
  cap_upper = 20,
  cap_lower = -30,
  years,
  frequency = "biennial",
  discard_rate = NA,
  ...
)
}
\arguments{
\item{object}{Optional. An object of class \code{rfb}.}

\item{A}{The reference catch. Should be an object of class \code{A}, see [A()].}

\item{r}{The biomass index ratio. Should be an object of class \code{r}, see [r()].}

\item{b}{The biomass safeguard. Should be an object of class \code{b}, see [b()].}

\item{m}{The multiplier. Should be an object of class \code{m}, see [m()].}

\item{cap}{\code{logical}. The uncertainty cap (stability clause). Defaults to \code{TRUE}}

\item{cap_upper}{Optional. \code{numeric}. The maximum allowed increase in the catch advice in \%. Default to +20.}

\item{cap_lower}{Optional. \code{numeric}. The maximum allowed decrease in the catch advice in \%. Default to -20.}

\item{years}{Optional. \code{numeric}. The years for which the advice should be given.}

\item{frequency}{Optional. The frequency of the advice ('annual'/'biennial'/'triennial'). Defaults to 'biennial'.}

\item{discard_rate}{Optional. The discard rate for the advice (\code{numeric}). If provided, advice values for catch and landings are given.}

\item{...}{Additional parameters. Not used.}
}
\value{
An object of class \code{rb}.
}
\description{
This function applies the rb rule.
}
\details{
The function requires the elements of the rb rule: A (the reference)
catch, r (the biomass index ratio), f (the fising pressure proxy), 
b (the biomass safeguard) and m (the multiplier). See the [A()], 
[r()], [b()], and [m()] help files for details.
}
\section{Warning}{

For application in ICES, do not change the default parameters (frequency, 
stability clause, etc) unless the changes are supported by case-specific
simulations.
}

\examples{
#' # calculate elements of rb rule for plaice
# reference catch
data(ple7e_catch)
A <- A(object = ple7e_catch, basis = "advice", units = "tonnes", advice_metric = "catch")
# biomass index trend
data(ple7e_idx)
r <- r(ple7e_idx)
# biomass safeguard
b <- b(ple7e_idx)
# multiplier
m <- m(hcr = "rb", k = 0.1)
# apply rb rule
advice <- rb(A = A, r = r, b = b, m = m, discard_rate = 27)
advice
advice(advice)

### application in subsequent years (without updating reference levels)
A <- A(object = ple7e_catch, basis = "advice", units = "tonnes", advice_metric = "catch")
r <- r(ple7e_idx)
b <- b(ple7e_idx, yr_ref = 2007) # use reference year for Itrigger
m <- m(0.95) # keep multiplier
advice <- rb(A = A, r = r, b = b, m = m, discard_rate = 27)
advice
advice(advice)

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Application of explicit precautionary principles in data-limited fisheries management. ICES Journal of Marine Science, 78: 2931--2942. \url{https://doi.org/10.1093/icesjms/fsab169}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2021. Using a genetic algorithm to optimize a data-limited catch rule. ICES Journal of Marine Science, 78: 1311--1323. \url{https://doi.org/10.1093/icesjms/fsab018}.

Fischer, S. H., De Oliveira, J. A. A., and Kell, L. T. 2020. Linking the performance of a data-limited empirical catch rule to life-history traits. ICES Journal of Marine Science, 77: 1914--1926. \url{https://doi.org/10.1093/icesjms/fsaa054}.
}
