% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfb.R
\docType{class}
\name{rfb-class}
\alias{rfb-class}
\title{An S4 class to represent the rfb rule.}
\description{
This class contains the components of the rfb rule (\code{rfb_A},
\code(rfb_r), \code(rfb_f), \code(rfb_b), \code(rfb_m))
}
\section{Slots}{

\describe{
\item{\code{advice}}{The value of the catch advice.}

\item{\code{advice_landings}}{Landings corresponding to the catch advice.}

\item{\code{advice_uncapped}}{The value of the catch advice without the uncertainty cap.}

\item{\code{units}}{The unit (e.g. tonnes) of the catch advice.}

\item{\code{advice_metric}}{The advice metric, 'catch' or 'landings'.}

\item{\code{frequency}}{The advice frequence (annual/biennial).}

\item{\code{years}}{The years for which the advice is valid.}

\item{\code{A}}{Component A (the reference catch).}

\item{\code{r}}{Component r (the biomass index ratio).}

\item{\code{f}}{Component f (the fishing pressure proxy).}

\item{\code{b}}{Component b (the biomass safeguard).}

\item{\code{m}}{Component m (the multiplier).}

\item{\code{cap}}{Uncertainty cap (restricts changes in advice).}

\item{\code{advice_change}}{Change in advice compared to previous advice.}

\item{\code{discard_rate}}{Discard rate (%).}
}}

