% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rfb_plot}
\alias{rfb_plot}
\alias{plot,r,missing-method}
\alias{plot,b,missing-method}
\alias{plot,r,b-method}
\alias{plot,b,r-method}
\alias{plot,I,missing-method}
\alias{plot,f,missing-method}
\alias{plot,F,missing-method}
\alias{plot,HR,ANY-method}
\title{Plot elements of the rfb/rb/chr rule}
\usage{
plot(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{r,missing}(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{b,missing}(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{r,b}(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{b,r}(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{I,missing}(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{f,missing}(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{F,missing}(x, y, y_label, show.data = TRUE, ...)

\S4method{plot}{HR,ANY}(x, y, y_label, show.data = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{rfb_r}, \code{rfb_b}, ...}

\item{y}{Optional. An additional object of \code{rfb_b}, ...}

\item{y_label}{Optional. The y-axis label.}

\item{show.data}{Show input data. Only applicable to some methods.}

\item{...}{Additional arguments. Not used.}
}
\value{
An object of class \code{gg}/\code{ggplot} with the plot.
Can be manipulated with the usual ggplot2 commands, e.g. \code{ylim()}.
}
\description{
A convenience function for plotting elements of the rfb, rb, and chr rules
using ggplot2 and loosely following ICES style figures.
}
\details{
Combinations of object are possible, e.g. for the rfb rule, it is possible to
plot components r (index ratio) and b (biomass safeguard) on the same plot.
}
\examples{
# Component b 
df_idx <- data.frame(year = 2017:2021,
                     index = c(1.33, 1.13, 0.84, 0.60, 1.03))
b <- rfb_b(df_idx)
plot(b, y_label = "Biomass index in kg/hr")

# Component r
r <- rfb_r(df_idx)
plot(r, y_label = "Biomass index in kg/hr")

# Components r and b combined
plot(r, b)

}
