% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/r.R
\name{show}
\alias{show}
\alias{show,r-method}
\title{show

Show a summary of the elements of the empirical harvest control rule}
\usage{
\S4method{show}{r}(object)
}
\description{
show

Show a summary of the elements of the empirical harvest control rule
}
