% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/A.R, R/length_data.R, R/F_hr.R,
%   R/b.R, R/m.R, R/r.R
\name{summary}
\alias{summary}
\alias{summary,A-method}
\alias{summary,Lc-method}
\alias{summary,Lmean-method}
\alias{summary,HR-method}
\alias{summary,b-method}
\alias{summary,m-method}
\alias{summary,r-method}
\title{summary}
\usage{
\S4method{summary}{A}(object, ...)

\S4method{summary}{HR}(object)

\S4method{summary}{b}(object, ...)

\S4method{summary}{m}(object, ...)

\S4method{summary}{r}(object, ...)
}
\arguments{
\item{object}{The object for which a summary is requested.}

\item{...}{Additional arguments. Not used.}
}
\description{
This function returns a summary of any component or the advice
for the empirical harvest control rules
}
