% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/comp_f.R, R/comp_m.R, R/comp_r.R
\name{value}
\alias{value}
\alias{value,comp_f-method}
\alias{value,comp_m-method}
\alias{value,comp_r-method}
\title{#' @rdname advice
#' @export
setMethod(f = "advice", signature = "comp_m", definition = function(object) 1)}
\usage{
value(object)

\S4method{value}{comp_f}(object)

\S4method{value}{comp_m}(object)

\S4method{value}{comp_r}(object)
}
\description{
#' @rdname advice
#' @export
setMethod(f = "advice", signature = "comp_m", definition = function(object) 1)
}
