% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_F_hr.R
\name{F}
\alias{F}
\alias{F,HR,comp_f-method}
\alias{F,HR,missing-method}
\alias{F,numeric,missing-method}
\alias{F,F,missing-method}
\title{Calculation of the (relative) harvest rate target}
\usage{
F(object, indicator, yr_ref, units, ...)
}
\arguments{
\item{object}{The time series with (relative) harvest rate values. See \code{\link{HR}}.}

\item{indicator}{The length based indicator. See \code{\link{comp_f}}.}

\item{yr_ref}{Optional. Allows direct specification of years to include in the calculation instead of using \code{indicator}.}

\item{units}{Optional. The units of the harvest rate. Can be derived automatically from argument \code{HR}.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{F} with the target harvest rate and the input data.
}
\description{
This function calculates the target harvest rate for chr rule.
}
\details{
Usually, this functions is used by providing a time series of (relative) harvest rate values (see \code{\link{HR}}) and a length-based indicator based on the mean catch length (see \code{\link{comp_f}}). The functions then finds those years where the indicator values is above 1, indicating that the fishing pressure is likely below Fmsy, extracts the corresponding (relative) harvest rate values for these years, and returns the average of these values as the target harvest rate. Alternatively, years can directly be specified with the argument \code{yr_ref} and the target harvest rate is then calculated as the average of the (relative) harvest rates for these years. See the ICES technical guidelines (ICES, 2022) for details.

If an object of class \code{F} is provided, its validity is checked.
}
\section{Warning}{

For application in ICES, the target harvest rate should only be calculated in the first year the chr rule is used and the same value used in subsequent years.
}

\examples{

# load harvest rate time series and length-based indicator
data(ple7e_hr)
data(ple7e_f2)
# calculate target harvest rate
F(ple7e_hr, ple7e_f2)

# use reference years when using in following years
F(ple7e_hr, yr_ref = c(2016, 2019))

}
\references{
ICES. 2022. ICES technical guidance for harvest control rules and stock assessments for stocks in categories 2 and 3. In Report of ICES Advisory Committee, 2022. ICES Advice 2022, Section 16.4.11, 20 pp. \url{https://doi.org/10.17895/ices.advice.19801564}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2022. Exploring a relative harvest rate strategy for moderately data-limited fisheries management. ICES Journal of Marine Science, 79: 1730--1741. \url{https://doi.org/10.1093/icesjms/fsac103}.
}
