% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_hr.R
\name{HR}
\alias{HR}
\alias{HR,data.frame-method}
\alias{HR,HR-method}
\title{Calculation of the (relative) harvest rate}
\usage{
HR(
  object,
  split_discards = FALSE,
  discard_survival,
  units_catch,
  units_index,
  units,
  ...
)
}
\arguments{
\item{object}{The data to use. Usually a \code{data.frame} with columns 'year', 'catch' and 'index'.}

\item{split_discards}{Shall the catch be split into landings and discards? Defaults to \code{FALSE}.}

\item{discard_survival}{Discard survival (\%). If \code{split_discards=TRUE}, this will be used to calculate the dead discards and these will be used in the harvest rate calculation.}

\item{units_catch}{Optional. The units of the catch, e.g. 'tonnes'.}

\item{units_index}{Optional. The units of the biomass index, e.g. 'kg/hr'.}

\item{units}{Optional. The units of the harvest rate. Can be derived automatically from \code{units_catch} and \code{units_index}.}

\item{...}{Additional arguments. Not currently used.}
}
\value{
An object of class \code{HR} with the harvest rate value(s).
}
\description{
The (relative) harvest rate is calculated by dividing the catch values by biomass index values.
}
\details{
Usually, this functions is used by providing a \code{data.frame} with columns 'year', 'catch' and 'index'. 
The catch can be split into landings and discards by providing 'landings' and 'discards' columns. 
The harvest rate can be calculated on the dead catch by specifying 
\code{split_catch=TRUE} and defining \code{discard_survival}. In this case,
the harvest rate will be calculated on the dead catch (landings plus
proportion of discards that die).

If an object of class \code{HR} is provided, its validity is checked.
}
\examples{

# create data.frame with data
df <- data.frame(year = 2017:2021, 
                 index = c(1.33, 1.13, 0.84, 0.60, 1.03),
                 catch = c(2949, 2513, 2091, 1888, 1615))
# calculate harvest rate
HR(df)

# check objects validity
HR(HR(df))

# plot harvest rate
# plot(HR(df, units_catch = "tonnes", units_index = "kg/hr"))

}
\references{
ICES. 2025. ICES Guidelines - Advice rules for stocks in category 2 and 3. Version 3. ICES Guidelines and Policies - Advice Technical Guidelines. 31 pp. \url{https://doi.org/10.17895/ices.pub.28506179}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2023. Risk equivalence in data‐limited and data‐rich fisheries management: An example based on the ICES advice framework. Fish and Fisheries, 24: 231--247. \url{https://doi.org/10.1111/faf.12722}.

Fischer, S. H., De Oliveira, J. A. A., Mumford, J. D., and Kell, L. T. 2022. Exploring a relative harvest rate strategy for moderately data-limited fisheries management. ICES Journal of Marine Science, 79: 1730--1741. \url{https://doi.org/10.1093/icesjms/fsac103}.
}
