% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/comp_A.R, R/comp_b.R,
%   R/length_data.R, R/comp_f.R, R/comp_m.R, R/comp_r.R
\name{summary}
\alias{summary}
\alias{summary,comp_A-method}
\alias{summary,comp_b-method}
\alias{summary,Lc-method}
\alias{summary,Lmean-method}
\alias{summary,comp_f-method}
\alias{summary,comp_m-method}
\alias{summary,comp_r-method}
\title{summary}
\usage{
summary(object, ...)

\S4method{summary}{comp_A}(object)

\S4method{summary}{comp_b}(object)

\S4method{summary}{comp_f}(object)

\S4method{summary}{comp_m}(object)

\S4method{summary}{comp_r}(object)
}
\arguments{
\item{object}{The object to be plotted.}

\item{...}{Additional arguments. Not used.}
}
\description{
This function returns a summary of any component or the advice
for the empirical harvest control rules
}
