% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/length_data.R, R/f.R, R/F_hr.R,
%   R/I.R, R/m.R, R/chr.R, R/r.R, R/rb.R, R/rfb.R
\name{value}
\alias{value}
\alias{value,Lc-method}
\alias{value,Lmean-method}
\alias{value,Lref-method}
\alias{value,f-method}
\alias{value,HR-method}
\alias{value,F-method}
\alias{value,I-method}
\alias{value,m-method}
\alias{value,chr-method}
\alias{value,r-method}
\alias{value,b-method}
\alias{value,rb-method}
\alias{value,rfb-method}
\title{value}
\usage{
value(object)

\S4method{value}{f}(object)

\S4method{value}{HR}(object)

\S4method{value}{F}(object)

\S4method{value}{I}(object)

\S4method{value}{m}(object)

\S4method{value}{chr}(object)

\S4method{value}{r}(object)

\S4method{value}{b}(object)

\S4method{value}{rb}(object)

\S4method{value}{rfb}(object)
}
\arguments{
\item{object}{The object from which the value is extracted.}
}
\description{
This function returns the value of any component or the advice
for the empirical harvest control rules
}
