% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{value}
\alias{value}
\title{#' @rdname advice
#' @export
setMethod(f = "advice", signature = "comp_m", definition = function(object) 1)}
\usage{
value(object)
}
\description{
#' @rdname advice
#' @export
setMethod(f = "advice", signature = "comp_m", definition = function(object) 1)
}
