% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/comp_I.R, R/length_data.R,
%   R/comp_f.R, R/comp_m.R, R/comp_r.R, R/rb.R, R/rfb.R
\name{value}
\alias{value}
\alias{value,comp_I-method}
\alias{value,Lc-method}
\alias{value,Lmean-method}
\alias{value,Lref-method}
\alias{value,comp_f-method}
\alias{value,comp_m-method}
\alias{value,comp_r-method}
\alias{value,comp_b-method}
\alias{value,rb-method}
\alias{value,rfb-method}
\title{value}
\usage{
value(object)

\S4method{value}{comp_I}(object)

\S4method{value}{comp_f}(object)

\S4method{value}{comp_m}(object)

\S4method{value}{comp_r}(object)

\S4method{value}{comp_b}(object)

\S4method{value}{rb}(object)

\S4method{value}{rfb}(object)
}
\arguments{
\item{object}{The object from which the value is extracted.}
}
\description{
This function returns the value of any component or the advice
for the empirical harvest control rules
}
